Added select serving amount 1-5 in recipe detail page

+++ backend/blog_api/views.py
@@ -743,17 +743,23 @@ def eaten_toggle(request):
         user = get_object_or_404(User, pk=user_id)
         blog_id = request.GET.get('blog_id')
         blog_ = get_object_or_404(blog, pk=blog_id)
-        serving = request.GET.get('serving')
+        serving = float(request.GET.get('serving'))
+        current_date = timezone.datetime.today().date()
 
-        user_eaten = Eaten.objects.filter(
-            userId=user, blogId=blog_, eaten_serving=serving)
+        # Check if an Eaten entry exists for the user, blog, and current date
+        user_eaten = Eaten.objects.filter(userId=user, blogId=blog_, eatenDate=current_date)
 
         if user_eaten.exists():
-            user_eaten.delete()
-            is_eaten = False
+            eaten_entry = user_eaten.first()
+            if eaten_entry.eaten_serving == serving:
+                user_eaten.delete()
+                is_eaten = False
+            else:
+                eaten_entry.eaten_serving = serving
+                eaten_entry.save()
+                is_eaten = True
         else:
-            Eaten.objects.create(userId=user, blogId=blog_,
-                                 eaten_serving=serving)
+            Eaten.objects.create(userId=user, blogId=blog_, eaten_serving=serving, eatenDate=current_date)
             is_eaten = True
 
         eaten_data = {
@@ -773,7 +779,7 @@ def eaten_exist(request):
     blog_ = get_object_or_404(blog, pk=blog_id)
     user_eaten = Eaten.objects.filter(userId=user, blogId=blog_)
     if user_eaten.exists():
-        return Response({"is_eaten": True})
+        return Response({"is_eaten": True, "serving": user_eaten.first().eaten_serving})
     return Response({"is_eaten": False})
